backend-app-api: extract and add tests for createSpecializedBackend
